+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for vk_lod_clusters
 
+* 2026-4-7:
+  * Bugfix to file cache header detection.
+  * The file cache can be used via memory mapping, avoiding a copy into system memory. `-mappedcache 0/1` defaults to true.
+  * Use "octant" encoding for vertex normals according to [A Survey of Efficient Representations for Independent Unit Vectors](http://jcgt.org/published/0003/02/01/paper.pdf)
 * 2025-4-4: 
   * The file cache format now stores everything geometry related for rendering. Instance and material information, as well as original vertex/triangle counts still comes from the gltf. The new file ending is `.nvsngeo`, the old `.nvcllod` files no longer work.
   * Added `-autoloadcache 0/1` option to disable loading from a cache file.